refactor(java): consolidate POM parser directly inside internal/librarian/java
This commit consolidates the POM Project unmarshalling helper files directly inside internal/librarian/java to resolve the final structural reviews:

- Deletes the root internal/maven/ package entirely.

- Creates unexported maven.go and maven_test.go directly inside internal/librarian/java.

- Keeps pomProject struct, ParsePOM() unexported as parsePOM(), and sentinel errors unexported in package scope to enforce Go encapsulation best practices.
